+++ test/unit/lib/securityHandlerTests.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2023 F5 Networks, Inc.
+ * Copyright 2023 F5, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -183,7 +183,6 @@ describe('SecurityHandler', () => {
             const securityHandler = new SecurityHandler(declaration, bigIpMock, undefined, state);
             return securityHandler.process()
                 .then(() => {
-                    console.log(dataSent);
                     const antivirusProtection = dataSent[PATHS.AntiVirusProtection];
                     const setting0 = dataSent[`${PATHS.WafAdvancedSettings}/Bo68NqP9roUE8Vv2NO-29Q`];
                     const setting1 = dataSent[`${PATHS.WafAdvancedSettings}/4NRiSGFR-qvXsN8VM7oKiw`];